updated import data script to be importable
