+++ src/utils/formatting/index.ts
@@ -1,0 +1,3 @@
+export function capitalize(word: string): string {
+  return word.charAt(0).toUpperCase() + word.slice(1);
+}
