+++ src/utils/types/models.ts
@@ -1,21 +1,20 @@
 import { z } from 'zod';
 import {
-  months,
   zipCodes,
   locations,
   unitedWayFocusAreas,
   focusAreaIndicators,
 } from '../constants';
 
-const monthsEnum = z.enum(months);
+// const monthsEnum = z.enum(months);
 const zipCodesEnum = z.enum(zipCodes);
 const locationsEnum = z.enum(locations);
 const focusAreasEnum = z.enum(unitedWayFocusAreas);
 
 const focusAreasIndicatorsEnum = z.enum(focusAreaIndicators);
 
-const TimePeriodSchema = z.object({
-  month: monthsEnum,
+export const TimePeriodSchema = z.object({
+  month: z.string(),
   year: z.string(),
 });
 
@@ -145,3 +144,4 @@ export type Report = z.infer<typeof ReportSchema>;
 export type TwelveMonthReport = z.infer<typeof TwelveMonthReportSchema>;
 export type Organization = z.infer<typeof OrganizationSchema>;
 export type Project = z.infer<typeof ProjectSchema>;
+export type TimePeriod = z.infer<typeof TimePeriodSchema>;
